requirement.txt
langchain==0.3.7
langchain-core==0.3.17
langchain-community==0.3.7

langchain-google-genai==2.0.6
google-generativeai==0.8.3

langgraph==0.2.36

pydantic==2.9.2
pydantic-settings==2.4.0

python-dotenv==1.0.1
